requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
graphviz==0.21
kiwisolver==1.5.0
matplotlib==3.10.9
neat-python==2.0.0
numpy==2.4.4
packaging==26.2
pillow==12.2.0
pygame==2.6.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
